This too

+++ src/Ruleset/RuleItem.cpp
@@ -57,7 +57,7 @@ void RuleItem::load(const YAML::Node &node, int modIndex, int listOrder)
 	_type = node["type"].as<std::string>(_type);
 	_name = node["name"].as<std::string>(_name);
 	_requires = node["requires"].as< std::vector<std::string> >(_requires);
-	_size = node["size"].as<float>(_size);
+	_size = node["size"].as<double>(_size);
 	_costBuy = node["costBuy"].as<int>(_costBuy);
 	_costSell = node["costSell"].as<int>(_costSell);
 	_transferTime = node["transferTime"].as<int>(_transferTime);
